% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phred.R
\name{check_phred}
\alias{check_phred}
\title{check_phred}
\usage{
check_phred(filenames, nlines = 1000)
}
\arguments{
\item{nlines}{number of quality lines to check.}

\item{filename}{name of fastq file(s) to check.}
}
\description{
checks the Phred encoding for the quality scores in fastq files.
}
