% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phred.R
\name{check_phred_range}
\alias{check_phred_range}
\title{check_phred_range}
\usage{
check_phred_range(filename, nlines = 1000)
}
\arguments{
\item{filename}{name of fastq file to check.}

\item{nlines}{number of quality lines to check.}
}
\description{
checks the Phred encoding for the quality scores in a fastq file.
}
