% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{checkAlphabetDistribution}
\alias{checkAlphabetDistribution}
\title{checkAlphabetDistribution}
\usage{
checkAlphabetDistribution(files)
}
\arguments{
\item{files}{character vector with the name of the files to process.}
}
\value{
a data.frame with the letter distribution.
}
\description{
Compute the distribution of letters in a Fastq file.
}
\examples{
NULL
}

