% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{fsize}
\alias{fsize}
\title{fsize}
\usage{
fsize(x, check = TRUE)
}
\arguments{
\item{x}{GenomicAlignmentPair object}

\item{check}{whether to check consistency of pair's position.}
}
\value{
an integer vector with fragment size.
}
\description{
Computes the fragment size.
}
\note{
Adapted from Herve Pages's implementation (TODO: add link to original).
}
\examples{
NULL
}
