% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phred.R
\name{get_bowtie2_phred}
\alias{get_bowtie2_phred}
\title{get_bowtie2_phred}
\usage{
get_bowtie2_phred(filenames, nlines = 1000)
}
\arguments{
\item{filenames}{character vector with the names of the files.}

\item{nlines}{number of files in the FASTQ file used to guess the phred.}
}
\description{
get_bowtie2_phred
}
