% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phastcons.R
\name{getPhastconsScores}
\alias{getPhastconsScores}
\title{getPhastconsScores}
\usage{
getPhastconsScores(db, chr, start, end)
}
\arguments{
\item{db}{the database of phastCons scores.}

\item{chr}{the chromosome.}

\item{start}{the starting point.}

\item{end}{the end point.}
}
\value{
an integer vector of phastCons scores.
}
\description{
Return phastCons scores.
}
\examples{
NULL
}
