% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{ngstools}
\alias{ngstools}
\alias{ngstools-package}
\title{Miscellaneous Tools for the Analysis of NGS Data}
\description{
For the moment provides some useful tools for the analysis of NGS data.
Helpful to keep things in one place.
}

