% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phastcons.R
\name{processPhastCons}
\alias{processPhastCons}
\title{processPhastCons}
\usage{
processPhastCons(object)
}
\arguments{
\item{object}{a BigWig file object.}
}
\value{
an RleList with the scores by chromosome.
}
\description{
Process a BigWig file with phastConst scores.
}
\examples{
NULL
}
